+++ libs/dot-components/src/lib/components/CommonProps.spec.tsx
@@ -1,0 +1,15 @@
+import { CommonProps } from './CommonProps';
+
+describe('CommonProps', () => {
+  it('should have unchanged API', () => {
+    const props = {
+      className: 'test-class',
+      'data-testid': 'testid',
+    };
+    const commonProps: CommonProps = {
+      className: 'test-class',
+      'data-testid': 'testid',
+    };
+    expect(commonProps).toEqual(props);
+  });
+});